Fixed edit to re-render
