+++ client/store/singleProduct.js
@@ -2,20 +2,12 @@ import axios from 'axios'
 
 //Action Types
 const GET_ITEM = `GET_ITEM`
-const UPDATE_ITEM = `UPDATE_ITEM`
 
 //ACtion CREATORS
 export const getItem = item => {
   return {
     type: GET_ITEM,
     item
-  }
-}
-
-export const updateItem = updatedItem => {
-  return {
-    type: UPDATE_ITEM,
-    updatedItem
   }
 }
 
@@ -31,28 +23,12 @@ export const getItemThunk = itemID => {
   }
 }
 
-export const updateItemThunk = (itemID, submittedChanges) => {
-  return async dispatch => {
-    try {
-      const {data} = await axios.put(
-        `/api/products/${itemID}`,
-        submittedChanges
-      )
-      dispatch(updateItem(data))
-    } catch (error) {
-      console.log(error)
-    }
-  }
-}
-
 const initState = {}
 
 export default function(state = initState, action) {
   switch (action.type) {
     case GET_ITEM:
       return action.item
-    case UPDATE_ITEM:
-      return action.updatedItem
     default:
       return state
   }
